requirements.txt
asgiref==3.2.3
Django==3.0b1
django-cors-headers==3.1.1
django-rest-base==0.1.6
djangorestframework==3.10.3
psycopg2-binary==2.8.4
uvicorn==0.10.4
